Remove generic-style implementation
