+++ CliToolkit.TestApp/Commands/ConfigurationValueCommand.cs
@@ -1,15 +1,10 @@
 ﻿namespace CliToolkit.TestApp.Commands
 {
-    public class ConfigurationValueCommand : CliCommand<ConfigurationValueOptions>
+    public class ConfigurationValueCommand : CliCommand
     {
-        protected override void OnExecute(ConfigurationValueOptions options, string[] args)
+        protected override void OnExecute(string[] args)
         {
             throw new System.NotImplementedException();
         }
-    }
-
-    public class ConfigurationValueOptions
-    {
-
     }
 }
