+++ CliToolkit.sln
@@ -15,6 +15,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CliToolkit.Tests", "CliToolkit.Tests\CliToolkit.Tests.csproj", "{7750A1CA-7C48-4E71-A81D-9C87CBD67FC6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -49,6 +51,18 @@ Global
 		{6BC57AD4-5C79-46EC-AE34-7C1CC647E7BB}.Release|x64.Build.0 = Release|Any CPU
 		{6BC57AD4-5C79-46EC-AE34-7C1CC647E7BB}.Release|x86.ActiveCfg = Release|Any CPU
 		{6BC57AD4-5C79-46EC-AE34-7C1CC647E7BB}.Release|x86.Build.0 = Release|Any CPU
+		{7750A1CA-7C48-4E71-A81D-9C87CBD67FC6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7750A1CA-7C48-4E71-A81D-9C87CBD67FC6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7750A1CA-7C48-4E71-A81D-9C87CBD67FC6}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{7750A1CA-7C48-4E71-A81D-9C87CBD67FC6}.Debug|x64.Build.0 = Debug|Any CPU
+		{7750A1CA-7C48-4E71-A81D-9C87CBD67FC6}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{7750A1CA-7C48-4E71-A81D-9C87CBD67FC6}.Debug|x86.Build.0 = Debug|Any CPU
+		{7750A1CA-7C48-4E71-A81D-9C87CBD67FC6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7750A1CA-7C48-4E71-A81D-9C87CBD67FC6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7750A1CA-7C48-4E71-A81D-9C87CBD67FC6}.Release|x64.ActiveCfg = Release|Any CPU
+		{7750A1CA-7C48-4E71-A81D-9C87CBD67FC6}.Release|x64.Build.0 = Release|Any CPU
+		{7750A1CA-7C48-4E71-A81D-9C87CBD67FC6}.Release|x86.ActiveCfg = Release|Any CPU
+		{7750A1CA-7C48-4E71-A81D-9C87CBD67FC6}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
